# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Helix API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/prompts/": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "List Prompts",
        "description": "List all registered prompts.",
        "operationId": "list_prompts_api_prompts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PromptSummary"
                  },
                  "type": "array",
                  "title": "Response List Prompts Api Prompts  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "prompts"
        ],
        "summary": "Create Prompt",
        "description": "Register a new prompt.",
        "operationId": "create_prompt_api_prompts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/extract-variables": {
      "post": {
        "tags": [
          "prompts"
        ],
        "summary": "Extract Variables",
        "description": "Extract Jinja2 template variables from a template string.\n\nUsed by the import flow to auto-detect variables before registration.",
        "operationId": "extract_variables_api_prompts_extract_variables_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractVariablesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractVariablesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/config": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Prompt Config",
        "description": "Return prompt config with provenance info.\n\nReads PromptConfig from DB. Returns both the effective (merged) config\nper role and the raw overrides dict, so the UI can show Global/Override badges.",
        "operationId": "get_prompt_config_api_prompts__prompt_id__config_get",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "prompts"
        ],
        "summary": "Update Prompt Config",
        "description": "Update per-prompt config in DB PromptConfig table.\n\nUpserts the PromptConfig row. Stores role-prefixed fields in the extra\nJSON column. Does NOT write config.json.",
        "operationId": "update_prompt_config_api_prompts__prompt_id__config_put",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptConfigSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Prompt",
        "description": "Get a prompt's full detail including template text from DB.",
        "operationId": "get_prompt_api_prompts__prompt_id__get",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/template": {
      "put": {
        "tags": [
          "prompts"
        ],
        "summary": "Update Template",
        "description": "Update a prompt's template text.",
        "operationId": "update_template_api_prompts__prompt_id__template_put",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/variable-definitions": {
      "put": {
        "tags": [
          "prompts"
        ],
        "summary": "Update Variable Definitions",
        "description": "Update variable definitions for a prompt.\n\nReplaces the stored variable definitions and re-derives anchor_variables.",
        "operationId": "update_variable_definitions_api_prompts__prompt_id__variable_definitions_put",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVariablesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/mocks": {
      "put": {
        "tags": [
          "prompts"
        ],
        "summary": "Update Mocks",
        "description": "Update mock definitions for a prompt.\n\nEach mock has tool_name + scenarios with match_args and response.",
        "operationId": "update_mocks_api_prompts__prompt_id__mocks_put",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMocksRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Mocks Api Prompts  Prompt Id  Mocks Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Mocks",
        "description": "Get mock definitions for a prompt.",
        "operationId": "get_mocks_api_prompts__prompt_id__mocks_get",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Mocks Api Prompts  Prompt Id  Mocks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/tools": {
      "put": {
        "tags": [
          "prompts"
        ],
        "summary": "Update Tools",
        "description": "Update tool definitions for a prompt.",
        "operationId": "update_tools_api_prompts__prompt_id__tools_put",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateToolsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Tools Api Prompts  Prompt Id  Tools Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExtractVariablesRequest": {
        "properties": {
          "template": {
            "type": "string",
            "title": "Template"
          }
        },
        "type": "object",
        "required": [
          "template"
        ],
        "title": "ExtractVariablesRequest",
        "description": "Request body for extracting variables from a template."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PromptSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "template_variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Template Variables"
          },
          "anchor_variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Anchor Variables"
          }
        },
        "type": "object",
        "required": [
          "id",
          "purpose",
          "template_variables",
          "anchor_variables"
        ],
        "title": "PromptSummary",
        "description": "Summary representation of a registered prompt."
      },
...
      "UpdateTemplateRequest": {
        "properties": {
          "template": {
            "type": "string",
            "title": "Template"
          }
        },
        "type": "object",
        "required": [
          "template"
        ],
        "title": "UpdateTemplateRequest",
        "description": "Request body for updating a prompt's template."
      },
...
    }
  }
}